Week 2: asking a question.
We've already know earlier that for Physics-based approach in graphics there are 2 options: Ray Tracing and Radiosity. We've already talked about RT and it's implementation in the consumer market (Nvidia's RTX line of GPUs).
But how about Radiosity? Is it not as popular as RT or just very difficult to implement to the consumer side. Because I haven't heard about any company implementing Radiosity on consumer-grade hardware.
Thank you Sir.