Could all browsers support a GCC compiler?
In the course, it is mentioned that we need to change the .c file into .o and compile it. But, we do know that compiling a .c file needs a GCC compiler. Could every browser compile it when its operating system still does not support GCC compiler?
Answer:
It is actually just a metaphor for the process that happens. It doesn't use a .c file. But it uses OpenGL ES Shader Language.